# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Bonus System API
  description: |
    API для управления бонусными программами с поддержкой мультитенантности и интеграцией с Telegram ботами.
    
    ## Основные возможности:
    - Управление проектами (тенантами)
    - Управление пользователями и бонусами
    - Webhook интеграция для внешних систем
    - Управление Telegram ботами
    - Аналитика и статистика
    - Реферальная программа
    - Многоуровневая система бонусов
  version: 1.0.0
  contact:
    email: support@saas-bonus.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.saas-bonus.com
    description: Production server
  - url: http://localhost:5006
    description: Development server
security:
  - JwtAuth: []
...
paths:
...
  /api/projects/{id}/users:
    get:
      tags:
        - Users
      summary: Получить список пользователей проекта
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: Поиск по имени, email или телефону
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Количество записей на странице (1-200)
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithBalance'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        telegramId:
          type: string
        telegramUsername:
          type: string
        isActive:
          type: boolean
        totalPurchases:
          type: number
        currentLevel:
          type: string
        referralCode:
          type: string
        referredBy:
          type: string
        registeredAt:
          type: string
          format: date-time
    UserWithBalance:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            currentBalance:
              type: number
            totalEarned:
              type: number
            totalSpent:
              type: number
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer